requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
gradio>=5.0
python-multipart==0.0.20
Pillow==11.2.1
opencv-python==4.11.0.86
numpy==2.2.4
PyMuPDF==1.25.5
rembg==2.0.65
onnxruntime==1.21.0
pytest==8.3.5
pydantic==2.11.3
